move utils to one place
