+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-allow-unreachable-code.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscAllowUnreachableCodeRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'allowUnreachableCode', expectedValue: false }),
