+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-allow-unused-labels.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscAllowUnusedLabelsRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'allowUnusedLabels', expectedValue: false }),
