+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-exact-optional-property-types.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscExactOptionalPropertyTypesRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'exactOptionalPropertyTypes', expectedValue: true }),
