+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-force-consistent-casing-in-file-names.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscForceConsistentCasingInFileNamesRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'forceConsistentCasingInFileNames', expectedValue: true }),
