+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-imports-not-used-as-values.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscImportsNotUsedAsValuesRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'importsNotUsedAsValues', expectedValue: 'remove' }),
