+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-no-implicit-returns.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscNoImplicitReturnsRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'noImplicitReturns', expectedValue: true }),
