+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-no-property-access-from-index-signature.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscNoPropertyAccessFromIndexSignatureRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'noPropertyAccessFromIndexSignature', expectedValue: true }),
