+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-no-unchecked-indexed-access.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscNoUncheckedIndexedAccessRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'noUncheckedIndexedAccess', expectedValue: true }),
