+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-no-unused-locals.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscNoUnusedLocalsRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'noUnusedLocals', expectedValue: true }),
