+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-no-unused-parameters.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscNoUnusedParametersRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'noUnusedParameters', expectedValue: true }),
