+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-strict.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscStrictRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'strict', expectedValue: true }),
