+++ packages/eslint-plugin-typescript/src/rules/ts-config/tsc-use-define-for-class-fields.rule.ts
@@ -1,5 +1,5 @@
 import { Rule } from 'eslint';
-import { tscRule } from './util';
+import { tscRule } from '../../util/tsc-util';
 
 export const TscUseDefineForClassFieldsRule: Rule.RuleModule = {
     create: (context) => tscRule(context, { key: 'useDefineForClassFields', expectedValue: true }),
