+++ packages/eslint-plugin-typescript/src/util/json-property-assertion.ts
