+++ packages/eslint-plugin-typescript/src/util/json-util.ts
